requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
gast==0.2.2
google-auth==1.17.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.2.1
pandas==1.0.5
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
python-dateutil==2.8.1
python-telegram-bot==12.8
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
tabulate==0.8.7
tensorboard==2.0.2
tensorflow==2.0.0
tensorflow-addons==0.9.1
tensorflow-estimator==2.0.1
termcolor==1.1.0
tornado==6.0.4
typeguard==2.9.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
